+++ src/main/webapp/i18/message_en.properties
@@ -30,6 +30,10 @@ modal.section.h5 = By signing up to create an account I accept Utility Bills Ter
 modal.section.h6 = Email
 modal.section.h7 = Password
 modal.section.h8 = Confirm Password
+modal.section.h9 = Authorisation Error.
+modal.section.h10 = Attention to the site entrance was not produced. You may have entered an incorrect username or password.
+modal.section.h111 = CONGRATULATION!!!
+modal.section.h112 = Now You Are On Our Team!
