+++ src/main/webapp/i18/message_ru.properties
@@ -29,6 +29,10 @@ modal.section.h5 = Создавая аккаунт, я принимаю усло
 modal.section.h6 = Электронный адрес
 modal.section.h7 = Пароль
 modal.section.h8 = Подтверждение пароля
+modal.section.h9 = Ошибка авторизации.
+modal.section.h10 = Внимание, вход на сайт не был произведен. Возможно, Вы ввели неверное имя пользователя или пароль.
+modal.section.h11 = ПОЗДРАВЛЯЕМ!!!
+modal.section.h12 = Теперь ты в нашей команде!
 
 
 #~~~~~~~~~~~About section~~~~~~~~~~~~~~~~~
